Add files via upload
Finding the mathematics table of any given number using do while statement.

A "do-while" loop in C++ is a control structure that allows you to repeatedly execute a block of code as long as a specified condition remains true. Unlike the "while" loop, which checks the condition before entering the loop, the "do-while" loop first executes the code block and then checks the condition.