Add files via upload
he std::swap function can be used for primitive data types such as int, float, etc., and also for user-defined data types as long as the appropriate operator= and copy constructors are defined for those types. Here's an example of how std::swap can be used
The std::swap function is a part of the Standard Template Library (STL) in C++, and it is generally efficient in swapping the values of variables. However, it is important to note that for user-defined types, the efficiency of std::swap heavily relies on the efficiency of the custom-defined operator= and copy constructors for the specific type.